requirements.txt
web3==7.14.1
python-dotenv==1.2.1
py-solc-x==2.0.5